api/requirements.txt
tensorflow>=2.21.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pillow>=10.0.0
matplotlib>=3.9.0
numpy>=1.26.0
requests>=2.32.0